NumericInput Component: Increment on empty field populates the field (#393)
**Pull Request Description**

Adjusts first click behavior so that value is rounded to nearest
increment.
When the current field is cleared the current field, then the increment
amount they click populates the field.
Refactors NumericInput component for cleaner implementation.

Closes #326 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Grant Braught <braught@dickinson.edu>
Co-authored-by: RolandLocke <roland.w.locke@gmail.com>
Co-authored-by: Grant Braught <braught@dickinson.edu>