fix(examples): fixes mixed type sorting in PickListBase (#408)
**Pull Request Description**

Fixes the sorting of mixed type columns (e.g.  numbers and strings).  

Also fixes example page issue related to `picked` prop not being
properly updated due to non trigger of reactivity and refactors
`handleSort`.

Closes #414 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: Grant Braught <braught@dickinson.edu>