feat(dev): added context sensitive button for PickerBase and PicklistBase (#402)
**Pull Request Description**

Adds context sensitive button for PickerBase and Picklist Base: See #166
and #186

Also added some styling on top of that logic to adjust for icon shift.

Closes #166
Closes #186 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Ty Chermsirivatana <chermsit@dickinson.edu>
Co-authored-by: Grant Braught <braught@dickinson.edu>