fix(dev): added persistent loop for authentication (#409)
**Pull Request Description**

Closes #407 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Ty Chermsirivatana <chermsit@dickinson.edu>
Co-authored-by: Grant Braught <braught@dickinson.edu>